Description
This python file, houses a class "Item". The purpose of the class is to store all of the relevant properties inside the object, and then access the properties that are stored inside itself. In this case we are requesting information from a website that does not produce dynamic numbers using javascript for the price of a desired item. We then store that page's HTML into an object, and then parse the HTML we stored. After we have parsed it, we then use Beautiful Soup to find our desired object we have defined in our init method, then formatting it so that we can store it into a float.