Item .py file
This python file contains the class Item. Where we request the HTML from a defined url and store it into an object. We then take the HTML parse it searching for defined terms in our init method. Once we have found the content we are looking for we manipulate the string so that we can store the item's price into a float.