Added app.py file to project
App.py will act as the heart of the program. Instead of defining our object's properties in the init method of the Item class in item.py, we will pass our object's desired properties through a constructor in main. This allows us to send a request for a different item that has different properties from within app.py.