Update item.py
In this commit, we have made it so that the init method for class Item can be called using a constructor from app.py. This allows us to call for different properties for a different object if we want to.

We have also made it so that self.price is defined in init, and load_price() returns self.price. So that when we call load_price method it will return our result, which is, the price of the item we are keeping our eye on.